Fix autoport issues
DEVSIX-662

+++ io/src/main/java/com/itextpdf/io/codec/TIFFDirectory.java
@@ -418,7 +418,7 @@ public class TIFFDirectory implements Serializable {
         if (i == null) {
             return null;
         } else {
-            return fields[i];
+            return fields[(int) i];
         }
     }
